испарвление ошибок и модули
