+++ index.html
@@ -6,9 +6,6 @@
   <link rel="stylesheet" href="css/normalize.css">
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
-  <script src="js/functions.js" defer></script>
-  <script src="js/main.js" defer></script>
-
   <title>Кекстаграм</title>
 </head>
 
@@ -236,6 +233,7 @@
         <h2 class="data-error__title">Не удалось загрузить данные</h2>
       </section>
     </template>
-
+    <script src="js/functions.js"  defer></script>
+    <script src="js/main.js" type="module" defer></script>
 </body>
 </html>
